replay: Rearrange modules
